ti: ln8000_charger_munch: Cast void pointer to int without warning
This fixes following warning
../drivers/power/supply/ti/bq2597x_charger_munch.c:2526:9: warning: cast to smaller integer type 'int' from 'const void *' [-Wvoid-pointer-to-int-cast]
        return (int)of_id->data;
               ^~~~~~~~~~~~~~~~
1 warning generated.

+++ drivers/power/supply/ti/ln8000_charger_munch.c
@@ -1652,7 +1652,7 @@ static int ln8000_get_dev_role(struct i2c_client *client)
 
 	dev_info(&client->dev, "%s: matched to %s\n", __func__, of_id->compatible);
 
-	return (int)of_id->data;
+	return (uintptr_t)of_id->data;
 }
 
 static int ln8000_parse_dt(struct ln8000_info *info)
